[#17] feature : assembler 구조 개편
