+++ src/components/assembler/AssembleFilePanel.tsx
@@ -1,20 +1,22 @@
-import {useRecoilState, useRecoilValue} from "recoil";
+import { useRecoilState, useRecoilValue } from "recoil";
 import {
-  assemblyExecutedLine,
   selectedAssemblyFileState,
   selectedFileContentState,
 } from "../../recoil/state";
 import Panel from "../common/Panel";
-import {useEffect} from "react";
-import {HL_ORANGE} from "../../styles/color";
+import { useEffect } from "react";
+import { HL_ORANGE } from "../../styles/color";
 import TopTab from "../common/TopTab";
 
-const AssembleFilePanel = () => {
+const AssembleFilePanel = ({
+  highlightNumbers,
+}: {
+  highlightNumbers: number[];
+}) => {
   const selectedAssemblyFile = useRecoilValue(selectedAssemblyFileState);
   const [fileContent, setFileContent] = useRecoilState(
     selectedFileContentState
   );
-  const highlightNumbers = useRecoilValue(assemblyExecutedLine);
 
   useEffect(() => {
     const fetchFile = async (filePath: string) => {
@@ -29,7 +31,7 @@ const AssembleFilePanel = () => {
   }, [selectedAssemblyFile, setFileContent]);
 
   return (
-    <div style={{flexDirection: "row"}}>
+    <div style={{ flexDirection: "row" }}>
       <TopTab title={selectedAssemblyFile} isBinary={false} />
       <Panel
         data={fileContent ? fileContent : []}
